[변경] Managers 폴더 변경
