+++ SOOUM/SOOUM/Data/Managers/AuthManager/Models/AuthInfo.swift
