+++ SOOUM/SOOUM/Data/Managers/GAManager/AnalyticsEventProtocol.swift
