+++ SOOUM/SOOUM/Data/Managers/GAManager/FirebaseLoggable.swift
