+++ SOOUM/SOOUM/Data/Managers/GAManager/GAManager.swift
