+++ SOOUM/SOOUM/Data/Managers/GAManager/SOMEvent.swift
