+++ SOOUM/SOOUM/Data/Managers/LocationManager/Models/Coordinate.swift
