+++ SOOUM/SOOUM/Data/Managers/ManagerConfiguration.swift
