+++ SOOUM/SOOUM/Data/Managers/ManagerType.swift
