+++ SOOUM/SOOUM/Data/Managers/NetworkManager/NetworkManager_Version.swift
@@ -1,0 +1,19 @@
+//
+//  NetworkManager_Version.swift
+//  SOOUM
+//
+//  Created by 오현식 on 9/16/25.
+//
+
+import Foundation
+
+import Alamofire
+import RxSwift
+
+
+// MARK: Version
+
+extension NetworkManager {
+    
+    
+}
