+++ SOOUM/SOOUM/Data/Managers/PushManager/Models/NotificationInfo.swift
